+++ moddb/utils.py
@@ -248,6 +248,90 @@ class Object:
     def __init__(self, **kwargs):
         self.__dict__.update(kwargs)
 
+def find(predicate, seq):
+    """A helper to return the first element found in the sequence
+    that meets the predicate. For example: ::
+
+        comment = find(lambda comment: comment.author.name == 'SilverElf', mod.comments.flatten())
+
+    would find the first :class:`.Comment` whose author's name is 'SilverElf' and return it.
+    If no entry is found, then ``None`` is returned.
+
+    This is different from `filter`_ due to the fact it stops the moment it finds
+    a valid entry.
+
+    .. _filter: https://docs.python.org/3.6/library/functions.html#filter
+
+    Parameters
+    -----------
+    predicate
+        A function that returns a boolean-like result.
+    seq : iterable
+        The iterable to search through.
+    """
+
+    for element in seq:
+        if predicate(element):
+            return element
+    return None
+
+def get(iterable, **attrs):
+    r"""A helper that returns the first element in the iterable that meets
+    all the traits passed in ``attrs``. This is an alternative for
+    :func:`moddb.utils.find`.
+
+    When multiple attributes are specified, they are checked using
+    logical AND, not logical OR. Meaning they have to meet every
+    attribute passed in and not one of them.
+
+    To have a nested attribute search (i.e. search by ``x.y``) then
+    pass in ``x__y`` as the keyword argument.
+
+    If nothing is found that matches the attributes passed, then
+    ``None`` is returned.
+
+    Examples
+    ---------
+
+    Basic usage:
+
+    .. code-block:: python3
+
+        article = moddb.utils.get(mod.get_articles(), name='Version 3.5 Released')
+
+    Multiple attribute matching:
+
+    .. code-block:: python3
+
+        comment = moddb.utils.get(mod.get_comments(2), content='Test', karma=3)
+
+    Nested attribute matching:
+
+    .. code-block:: python3
+
+        comment = moddb.utils.get(article.get_comments(), author__name='SilverElf', content='Best article ever')
+
+    Parameters
+    -----------
+    iterable
+        An iterable to search through.
+    \*\*attrs
+        Keyword arguments that denote attributes to search with.
+    """
+
+    def predicate(elem):
+        for attr, val in attrs.items():
+            nested = attr.split('__')
+            obj = elem
+            for attribute in nested:
+                obj = getattr(obj, attribute)
+
+            if obj != val:
+                return False
+        return True
+
+    return find(predicate, iterable)
+
 time_mapping = {
     "year" : 125798400,
     "month": 2419200,